+++ src/api/eval.rs
@@ -67,7 +67,7 @@ impl Engine {
         let ast = self.compile_with_scope_and_optimization_level(
             scope,
             &[script],
-            self.optimization_level,
+            self.options.optimization_level,
         )?;
         self.eval_ast_with_scope(scope, &ast)
     }
@@ -227,9 +227,9 @@ impl Engine {
             ast.as_ref(),
         ];
         let lib = if lib.first().map(|m: &&Module| m.is_empty()).unwrap_or(true) {
-            &lib[0..0]
+            &[]
         } else {
-            &lib
+            &lib[..]
         };
         self.eval_global_statements(scope, global, &mut state, statements, lib, level)
     }